move RenameClassConstantsUseToStringsRector from Nette to generic Rector/Constant
