--- packages/Nette/tests/Rector/Bootstrap/RenameClassConstantsUseToStringsRector/config.yml
@@ -1,2 +1,0 @@
-services:
-    Rector\Nette\Rector\Bootstrap\RenameClassConstantsUseToStringsRector: ~
