+++ tests/Rector/Constant/RenameClassConstantsUseToStringsRector/Correct/correct.php.inc
