+++ tests/Rector/Constant/RenameClassConstantsUseToStringsRector/RenameClassConstantsUseToStringsRectorTest.php
@@ -1,12 +1,12 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Nette\Tests\Rector\Bootstrap\RenameClassConstantsUseToStringsRector;
+namespace Rector\Tests\Rector\Constant\RenameClassConstantsUseToStringsRector;
 
 use Iterator;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
 
 /**
- * @covers \Rector\Nette\Rector\Bootstrap\RenameClassConstantsUseToStringsRector
+ * @covers \Rector\Rector\Constant\RenameClassConstantsUseToStringsRector
  */
 final class RenameClassConstantsUseToStringsRectorTest extends AbstractRectorTestCase
 {
