+++ tests/Rector/Constant/RenameClassConstantsUseToStringsRector/Wrong/wrong.php.inc
