+++ tests/Rector/Constant/RenameClassConstantsUseToStringsRector/config.yml
@@ -1,0 +1,2 @@
+services:
+    Rector\Rector\Constant\RenameClassConstantsUseToStringsRector: ~
